Convert CLI scripts to entry points
The new names for old tools are:

  * snmpsimd.py -> snmpsim-command-responder
  * snmprec.py -> snmpsim-record-commands
  * mib2dev.py -> snmpsim-record-mibs
  * pcap2dev.py -> snmpsim-record-traffic
  * datafile.py -> snmpsim-manage-records
